# perf: parallel Merkle tree construction