--- a/crypto/Cargo.toml
+++ b/crypto/Cargo.toml
@@ -13,6 +13,7 @@ sha3 = "0.10"
 sha2 = "0.10"
 thiserror = "1.0.38"
 serde = { version = "1.0", features = ["derive"] }
+rayon = { version = "1.8.0", optional = true }
 
 [dev-dependencies]
 criterion = "0.4"
@@ -21,6 +22,7 @@ rand = "0.8.5"
 
 [features]
 test_fiat_shamir = []
+parallel = ["dep:rayon"]
 
 [[bench]]
 name = "criterion_merkle"