--- a/crypto/src/merkle_tree/traits.rs
+++ b/crypto/src/merkle_tree/traits.rs
@@ -1,3 +1,6 @@
+#[cfg(feature = "parallel")]
+use rayon::prelude::{IntoParallelRefIterator, ParallelIterator};
+
 /// A backend for Merkle trees. This defines raw `Data` from which the Merkle
 /// tree is built from. It also defines the `Node` type and the hash function
 /// used to build parent nodes from children nodes.
@@ -11,10 +14,12 @@ pub trait IsMerkleTreeBackend: Default {
     /// This function takes the list of data from which the Merkle
     /// tree will be built from and converts it to a list of leaf nodes.
     fn hash_leaves(unhashed_leaves: &[Self::Data]) -> Vec<Self::Node> {
-        unhashed_leaves
-            .iter()
-            .map(|leaf| Self::hash_data(leaf))
-            .collect()
+        #[cfg(feature = "parallel")]
+        let iter = unhashed_leaves.par_iter();
+        #[cfg(not(feature = "parallel"))]
+        let iter = unhashed_leaves.iter();
+
+        iter.map(|leaf| Self::hash_data(leaf)).collect()
     }
 
     /// This function takes to children nodes and builds a new parent node.